Add Configure TypeScript in the project
