--- src/redux/StarRating/action.types.js
@@ -1,5 +1,0 @@
-const UserActionType ={
-    ATUALIZAR_STAR: 'StarRating/EnviarStar'
-}
-
-export default UserActionType;
