+++ src/redux/store.js
@@ -1,9 +1,7 @@
-import {applyMiddleware, createStore} from 'redux';
-import rootReducer from './root-reducer';
-import logger from 'redux-logger';
-
+import { applyMiddleware, createStore } from "redux";
+import rootReducer from "./root-reducer";
+import logger from "redux-logger";
 
 const store = createStore(rootReducer, applyMiddleware(logger));
-
 
 export default store;
